Pass child components into HealthCard to decouple HealthCard from VaccineCard and LabCard
